# perf(assump) : enhance get_known_facts_keys