… sympy/assumptions/ask_generated.py