<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->


#### Brief description of what is fixed or changed

When predicate decomposition was introduced, `get_known_facts_keys()` was modified to exclude the composite predicates. But this turned out to be not only unnecessary but slightly slows down the evaluation. This change reverts that.

#### Other comments

`get_known_facts_keys()` is involved in simple, limited case where a single unary predicate is passed as assumption, e.g. `ask(Q.real(x), Q.positive(x))`. That's why the regression was unnoticed.

In master:
```python
In [1]: %timeit ask(Q.real(x), Q.positive(x))
2.01 ms ± 5.94 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

In this branch:
```python
In [1]: %timeit ask(Q.real(x), Q.positive(x))
601 µs ± 2.16 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```


#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
NO ENTRY
<!-- END RELEASE NOTES -->